fix: crashing on iPad
for some reason, objectionable C decides that `playsinline` which is declared as an `NSNumber*`,
is actually an `NSTaggedPointerString`, which does not have the `isEqualToNumber` method that we're trying to use here.
Since we want to allow inline media playback on iPad anyways, I'm just going to comment this out and call it a day.